Improve testing of Singularity on EP startup. HTCONDOR-1431
Specifically, Singularity testing at startup should use the same
containment arguments that are used when actually running jobs.
While we are at it, improve SINGULARITY_USE_PID_NAMESPACES knob to have
and "auto" value by default, telling HTCSS to use pid namespaces if
that works, and fallback to not using them if it does not.