HTCONDOR-1448 Check new ComputingManagerEndTime in ARC job info data
In newer ARC CE versions, EndTime isn't set in info data for FAILED jobs.
The new ComputingManagerEndTime is present for FAILED jobs. Consider the
data non-stale if either element is present.
Before, the gridmanager ended up in an endless loop of retrying the info
query because EndTime was absent for FAILED jobs.