Introduce shadow prepare job hook.  HTCONDOR-1453
This introduces the shadow equivalent of the prepare job hook; the
shadow launches the hook prior to activating the job, allowing the
admin to do arbitrary transforms (such as staging input files and
then changing the TransferInput list).

Since the primary use case is file transfers - and that often requires
credentials - we make sure we stage the OAuth2 credentials so they
are available to the script.