HTCONDOR-1360 Improve starter handling of syscall socket errors
When a communications error occurs on the syscall socket (i.e. a CEDAR
call fails), the starter now always closes the connection and goes into
reconnect mode. Before, if the starter timed out waiting for a reply from
the shadow, it would often continue on and make addtional RPCs. If the
shadow eventually responded, the starter would interpret the reply for the
wrong RPC, confusing it.