HTCONDOR-1360 Improve shadow handling of syscall socket errors
If the shadow blocks for long enough, the starter can timeout, close the
syscall socket, and go into reconnect mode. When the shadow unblocks, it
will read the RPC and then fail to send a reply. Before, shadow would
ASSERT() and die in reaction. Now, it will close the syscall socket and go
into reconnect mode.

For now, just change the handling of end_of_message(). This is the call
that will fail in most cases when the connection has closed, and returning
on earlier CEDAR call failures will leak resources without more code
changes.