HTCONDOR-1463 Add SciTokens mapping plugin interface
Allow admin to provide plugins (external programs) for mapping a validated
SciToken to an HTCondor identity. Each plugin can provide a mapping or
decline to do so. In the latter case, additional plugins are consulted.
If no plugin produces a mapping, then the default mapping of
scitokens@unmapped is used. The plugin interface is intended to be
compatible with that supported by ARC CE.

If the mapped identity for SciTokens in the configured mapfile starts with
'PLUGIN:', then the rest of the identity is a comma-separated list of
plugin names to execute to produce the actual identity.

The config parameter SEC_SCITOKENS_PLUGIN_<name>_COMMAND provides the
executable and command-line arguments to run. The plugin is given the
token's payload via stdin and via environment variables as ARC CE provides
them (with names that start with 'BEARER_TOKEN_0'). If the plugin exits 0,
then it accepts the token. If the plugin doesn't print the mapped identity
in stdout, then config param SEC_SCITOKENS_PLUGIN_<name>_MAPPING must be
set to the mapped identity. If the plugin exits 1, then it doesn't accept
the token and additional plugins will be executed.

The special value 'PLUGIN:*' indicates that the config parameter
SEC_SCITOKENS_PLUGIN_NAMES should be used to provide the list of plugin
names. This is the default behavior if a mapfile isn't defined.

If a plugin doesn't execute properly (can't be spawned or exits with an
unexpected status), then authentication fails and the client is refused.