HTCONDOR-1498 Allow use of EGI CheckIn tokens for SciTokens auth
The SciTokens library mostly accepts an EGI CheckIn token when treated as
a SciTokens v1.0 token. The main issue is that the Enforcer objects to the
presence of claims in the token that aren't a part of SciTokens 1.0.
Add knob SEC_SCITOKENS_ALLOW_FOREIGN_TOKEN_TYPES that allows the SciTokens
Enforcer check to fail, with a default to not tolerate such failures.
For "foreign" tokens, search the full set of scopes for HTCondor
authorization-related scopes.