HTCONDOR-1491 Make cmd_status-* tests more robust
They stand up 2 personal condors, one of which has a schedd
but no negotiator (on purpose).  Rarely, submitting to that schedd
fails because the schedd is blocked, trying to send a reschedule command
from a prior submit to the negotiator.  Fix this in the test by
setting SCHEDD_SEND_RESCHEDULE=false for that schedd.