HTCONDOR-1577 remove "using namespace std" from classads
"using namespace std" is a code smell, as when newer versions
of c++ add new symbols to the std namespace, it leads to
potential namespace collisions.

I'm too lazy to replace all the unqualified symbols from the
std namespace with std:: prefixes, so we'll just import
the symbols we use.