HTCONDOR-1601 Allow use of C++20 in Windows
MSVC's C++20 mode is much more strict about C++ than C++17
It is much more const-correct, which is a headache to fix,
but generally a good thing to have.

It is so strict, that it can't compile Win7 headers, unless
you predeclare a "struct IUnknown" before include'ing
windows.h.  This patch does some of that, but not enough.
Do we really need to support Win7?

Because of that, we don't turn on C++20 mode in the Cmake
in this patch.  Nevertheless, the drudge work is done here.