Test transfer plugins if URL is specified.  HTCONDOR-1608
With this, the HTCondor file transfer plugins will be tested prior to
being inserted into the list of available URL schemas.

For example, for the HTTPS plugin, an admin might specify

```
HTTPS_TEST_URL = https://google.com
```

In this case, HTTPS will only be available if google.com is present
at startup.