HTCONDOR-1583 Fix stale data check for ARC CE jobs that failed at submit
Not all failed ARC CE jobs get an EndTime or ComputingManagerEndTime
element in their info data, for example if submission to the LRMS fails.
A more reliable indicator of non-stale data is if the State element
contains an item of the form "arcrest:XXX", where XXX is one of the
terminal states that triggered our query.