HTCONDOR-1695 Make c-gahp and job router set cluster attributes
The c-gahp and job router now put most job attributes in the cluster ad,
just like condor_submit and the python bindings do. This is in preparation
for some upcoming features in the schedd that will require come attributes
to appear in the cluster ad and not the proc ad.