HTCONDOR-1689: Default condor home to /var/lib/condor on Debian/Unbuntu
This commit corrects my blunder with confusing LOCAL_DIR and condor's
home directory. (HTCONDOR-1537) It retains the change to open up the
permissions on /var/lib/condor if it exists. This is required on later
versions of Debian and Ubuntu because user's home directories are not
world readble.  It conditionally checks for the directory's existance,
because if the script produces an error, the whole installation fails.