HTCONDOR-1757 Set SciTokens cache directory by default
Add config param SEC_SCITOKENS_CACHE, which specifies where we instruct
the SciTokens library to put its credential cache. Otherwise, it won't use
a cache if we don't have a home directory. The default  of "auto" means to
use $(RUN)/cache if RUN is defined, else use $(LOCK)/cache. This allows us
to use the best location for both linux native package and tarball
installations.
We are leaving this param undocumented for now.
All of this requires a new enough version of the SciTokens library, but we
need to work with older versions as well.