HTCONDOR-1310 Return claimed slot info when requesting a claim
In the REQUEST_CLAIM command, the schedd can now ask the startd to include
the machine ad and claim id of the claimed slot in its reply. This is
particularly useful when a new dynamic slot is being created. Before, the
schedd had to guess what the new slot ad would look like.