HTCONDOR-1801 fix cgroup OOM detection on el7
EL7 doesn't have the oom_kill field in the memory.oom_control
control file.  We still want to poll after job exit for OOM
killing, to avoid race conditions and tightly coupling the
cgroup code with the DaemonCore event loop.

So, this patch changes the OOM detection to create an eventfd,
tell the kernel to generate an event on it when the OOM killer
fires in a cgroup.  But, instead of putting this fd into the
daemon core select loop, we merely poll it after job exit
to see if the OOM killer fired.

This way, we are completely synchronous, and the cgroup
code is independent of the daemon core event loop.