HTCONDOR-1824 Allow claiming of partitionable slots
A pslot can now be claimed by the schedd (i.e. the pslot's State changes
to Claimed). This is done instead of claiming a new dslot under the pslot.
In the future, the schedd may be able to do both in a single
CLAIM_REQUEST.
Pslot claiming follows the model of the existing WorkingCM (with a set
lifetime for the claim and no preemption), but can be done with or without
requesting the startd to advertise to an additional "working" collector.

This also lays some groundwork for requesting a group of dslots to be
created in a single request.

This is only changes in the startd and DCStartd. Code to utilize this new
feature will follow.