HTCONDOR-1824 Schedd and negotiator now use claimed pslots
The schedd can now claim a pslot (vs claiming a new dslot under the
pslot). This is disabled by default (knob CLAIM_PARTITIONABLE_SLOT) and
doesn't happen if either the startd or the negotiator doesn't support it.
Pslot claiming works with SCHEDD_LOCAL_CM, but can be used without it.

The negotiator will now match claimed pslots, but will only do so for jobs
coming from the same schedd that has the pslot claimed. The negotiator can
be configured to not match claimed pslots if they're reporting to a
secondary collector (i.e. if WorkingCM is set in the slot ad).