HTCONDOR-1853 Fix intermittently failing job_core_niceuser_van
The test job_core_niceuser_van fails intermittently.  It starts a one
slot startd, and in one job cluster, submits two nice user jobs and two
non-nice user jobs, and tests that both non-nice user jobs start and run
first.

The problem is that when the schedd sends the two submitter ads to the
collector, it doesn’t do this atomically.  Rarely, the negotiator will
run when the nice user submitter ad is in the collector, but the
non-nice submitter ad is not.  In this case, we run the jobs in the
“wrong” order.

Ideally, we should have an interface to the collector where we can
atomically update multiple ads, but we don’t.  To fix this test, we will
turn off the negotiator, submit the jobs, and only after the submit
event is written, start up the negotiator again.