HTCONDOR-1871 fix race condition in test_checkpoint_destination
We see intermittent failures in the test_checkpoint_destination test.
Turns out this is because we run cp -a . {save_directory} in the execute
scratch dir, and this cp command occasionally fails with

/bin/cp: cannot stat './.update.ad.tmp': No such file or directory

This is because there is a race condition with the startd which writes
to .update.ad.tmp, and renames it, and cp -a is a two pass algorithm.
We then see unexpected output in stderr, which breaks the test.  Let’s
fix the test by just redirecting the stderr of the cp command to
/dev/null