HTCONDOR-1882: Setup for Initial LotMan Integration
Currently, the spool directory remains largely untracked, and we have little insight
as to how much different jobs/users are being allocated on it. LotMan is a data
accountant that can track usage over "lots", ie the entities to which data are assigned.
In this first-pass integration, the uid assigned to each job is being used as its "lot" name.

While the longterm goal is to track total usage of the spool with the ability to make
decisions about jobs based on current statistics (eg put a job on hold if the user is
already using too much spool space), this commit only encompasses tracking without the
decision making. For now, LotMan plugs into the FileTransfer object's DoDownload method,
which doesn't make file sizes available until after the download is complete. For decision
making, we have two options: find either a different place to tap into before files are
transferred that can access the to-be-downloaded files names (for determining if the files
are headed to spool), as well as their sizes, count, and associated UID, OR modify the FTO
to be aware of what it should expect as to file count and sizes before the download is
initiated.