HTCONDOR-1824 Add option to disable matching of claimed pslots
When claiming a pslot, the schedd can now request that no negotiators do
any matching of the slot (new attribute WantMatching=false is added to the
slot ad). The schedd doesn't use this feature yet.
Also fixed some bugs with managing WorkingCM and pslot lease time in the
startd and filtering of pslot ads in the negotiator.