"HTCONDOR-1858: Changed logic for test hook spawned by shadow"
The original implementation ran ps with many arguments to show the
parent of the shadow hook is the condor shadow. This created
issues shown in batlab with many different machines that either do not
have ps as a command or have ps but do not have the arguments initially
used. The new approach is to set a schedd environment variable and use
the shadow hook to see if it can access it. In addition, also used a
starter hook to ensure that it cannot access it (since it is on the
EP/starter). Also changed some comments in the file as well.