HTCONDOR-1914: fixed bug with shadow job hooks in rootly condor
Before this fix, shadow hooks running in a rootly condor would have an
issue with permissions with files in the procd: procd_pipe.watchdog and
procd_pipe. After conversations with Greg and Cole, we decided it is
best for shadow hooks to not go through the procd because there could be
too many connections with many shadow hooks interacting with 1 procd.
The fix checks if the hook is shadow or starter. If it is a starter
hook, then it will interact with the procd to keep track of any child
processes. If it is a shadow hook, it will not interact with the procd
and it should be cleaned up by the condor_shadow.