HTCONDOR-1925 Robustify cmd_status-compact test
This test submits a job, polls condor_status -schedd until
the JobsRunning attribute is non-zero, and analyzes the
output of condor_status -schedd.

Then it removes the job

In a subsequent phase, it also submits a job, polls condor_status
-schedd
until the the schedd ad reports nonzero JobsRunning, then runs
condor_status again to verify.

There is a race condition here, where the first phase succeeds,
and the test removes the job, but the schedd ad hasn't made it
to the collector yet.  We submit the second job, and though it
hasn't started, our polling of condor_status -schedd succeeds
with the stale info.  Between that time and the next run of
condor_status, the schedd updates its ad with 0 jobs running to
the collector, and hasn't gotten around to starting the second job

In this case the test fails with inconsistent data.

We fix this by not removing the first job, or submitting a second
job, leaving the first job running in the queue for the second
phase of the test.