HTCONDOR-1966 Fix credential service names in CONDOR_getcreds() RPC
The shadow used to send the filename that should be created as the service
name for each credential (replace '*' with '_' and append '.use').
Recently, the starter was changed to expect the '.use' but not the
substitution of '*'.
Now, the unmodified service name is sent for each credential. Both sides
can handle older peers that have the old behavior.
Also, the starter no longer EXCEPT()s if it sees unexpected data from the
shadow.