Do not start jobs with transfer queue over threshold. HTCONDOR-2048
When the transfer queue is above a given threshold, do not start
the corresponding job.  This mitigates situations where a user's
transfer queue is hopelessly overloaded - we should back off instead
of piling on more work.

The majority of this commit is around an improved heuristic for defining
an "overloaded" queue.  We now sum up the total transfers (queued plus
active) and divided it by the estimated bandwidth to get an estimate
for remaining transfer time; the threshold is the remaining transfer time
in seconds (default: 30 minutes).  If we are just starting from a cold
start, instead of relying on the bandwidth estimate, we assume a minimum
of 1MB/s.

In order to make `TRANSFER_QUEUE_USER_EXPR` available at match time, we
use a match ad including the slot ad; this way, one can set:

```
TRANSFER_QUEUE_USER_EXPR = SLOT.GLIDEIN_Site
```

(previously, the match attrs were available; the above syntax replaces
their use).  In the scheme above, if a particular site has significantly
less bandwidth than others, the schedd would stop accepting new matches
for it.