HTCONDOR-2072 Fix flakey test
job_rank_and_aclustering tries to rank jobs to run on certain
slots.  However, before doing that, it runs a random single job
for no apparent reason.  Rarely, after the first job exits, the startd
update hasn't made it to the collector, and no job matches
the slot that the first job ran it, throwing off the attempt
to force specific jobs to specific slots.  So, just don't
submit the first, useless job.