Add ability to adjust lotman db timeouts in WAL mode via LOTMAN_DB_TIMEOUT
With the latest version of LotMan, the sqlite database has been configured to run in WAL
mode, which increases the overall concurrency we can achieve. When concurrency is very high,
the database may slow down to a degree. By adjusting the timeout, we can tune for particular
systems. Note that in the event a timeout is reached, LotMan shouldn't block condor --
instead the db will just return a "busy" error, and the information won't be logged. While
this is bad for overall tracking, it's vital that Condor not be bricked by a slow database
because of LotMan.