HTCONDOR-2013 Use htcondor.log for credmon logging
While the Python logging library is useful, getting it to behave
like HTCondor's logging functionality is challenging. This PR
un-reinvents the wheel and uses the htcondor.log function provided
by the Python bindings inplace of calls to the Python logging library,
importantly only *after* the credmons have set their subsystem name
so that logs are not placed in ToolLog.

One drawback is that the OAuth webserver does not know what to set
its subsystem name to, so instead of using htcondor.log or the Python
logging library, it now emits log messages to stdout/stderr, which is
helpfully emitted into the webserver error logs, which is where these
types of message belong anyway.

Some very minor code cleanup items were also done. Main item here is
putting argument parsing into its own function to be shared by any
and all credmons (right now the OAuth and Vault credmons). More cleanup
is desired but would not be fit for a bugfix commit.

Note that while we continue to support EL7-based systems, f-strings
cannot be used in any code touched by the OAuth webserver, which
includes the "utils" library here.