HTCONDOR-2096 Add <service>_oauth_options to condor_submit
This commands allows the passing of extra options to the oauth credential
storer command, whose meaning is specific to a given command.
condor_submit treats the value as an opaque string.
Like the other oauth credential parameters, the config knobs
<service>_USER_DEFINE_OPTIONS and <service>_DEFAULT_OPTIONS control
whether the options command must be provided or provide a default value.