HTCONDOR-2173 Speed up ctest runs
ps shows that when I run ctest -j 40, there's usually a dozen
python processes running box|gdrive|onedrive plugin -classad.

Testing standalone shows that just getting python off the ground
takes about .2 of a second, and when we do three of these for
each startd, it adds up.  We have no tests that test these plugins,
and I'm not sure we have any users, so I'm just going to disable
them in ctest.

Ideally, I'd like the equivalent of "-=" in config language,
but for now, we'll hard code the transfer plugin list to just
include the curl plugin, which we do test, and is in native
condor code, so it is fast to run -classad.