HTCONDOR-2175 Add docs for local SciTokens issuer
Adds a section to the admin manual for how to set up the local issuer,
including the steps to generate a key pair and properly host the public
key. Also includes a job transform to automatically generate tokens on
submit.

The example credmon config file was also updated to better follow along
with the example in the documentation.