HTCONDOR-2236 fix bug in user log reader with short reads.
In the test suites, we (rarely) see dagman hanging, having lost track of
a job it submitted.  Turns out that the user log read code is to blame.
There are cases where even though the user log writer is writing the
entire event in one write(2) call, on an FD opened with O_APPEND, posix
only guarantees that a reader will see every byte of that write at once
if the read is started after the write returns to userspace. We make no
attempts to do so.

In a couple of places in the log reader code, we fseek back to the
beginning  when we detect a short read, but not all cases.  This PR adds
fseek’ing back to the remaining cases.