HTCONDOR-2292 Do normal shadow shutdown logic if job_exit() reply fails
If the syscall socket closes before the shadow sends its reply to the
job_exit() RPC, handle it the same as when the socket closes after
sending the reply (the usual case).
Otherwise, the shadow goes into reconnect mode and eventually exits
signaling the job should leave the queue without evaluating on-exit policy,
writing a terminate event, or sending a terminate email.