Remove uneeded spawning of ganglia tools.  HTCONDOR-2346
No need to popen gstat command if sending data for all hosts, and
also no need to spawn gmetric to send heartbeats if running in
NOOP mode.  Besides wasting cycles, a lot of sites do not have these
command-line tools installed, resulting in tons of spam in the log files.