HTCONDOR-2335 port to C++ 23
C++23 adds a deleted constructor to std::string:

std::string(nullptr_t) = delete;

This is wellness and good, because it turns a runtime crash into
a compile time error.  However, it also cause our HashTable Iterator
class to fail to compile.  This is in the Iterator dereference
functions:

std::pair<Index, Value> operator *() const {
   return std::pair<Index, Value>(m_cur ? m_cur->index : NULL, m_cur ? m_cur->value : NULL);
}

m_cur is null when the iterator is at end().  Technically, it is UB
to dereference an at-end iterator, but I guess we are trying to be
nice here.  However, if either the Index or the Value is of type
std::string, we try to assign NULL to the std::string, which triggers
the compile error.

I'm going to make the smallest possible change I can think of to
get this to compile, by default-initializing the index and value
returned.  If they are pointers, they will be nullptr.  If
the are strings they will be empty.

I believe it is technically correct to not check the validity
of m_cur here, and let it fault at runtime if we try to
dereference an at-end() iterator, but I'm not willing to make
that kind of breaking change just yet