HTCONDOR-2373 Store the exit code of file xfer plugin in hold_subcode
In 23.0, we changed the hold_subcode to contain 1 if the file xfer
plugin failed for any reason, no matter which non-zero exit code it
returned.  This PR changes the functionality back to the original,
where if the plugin exits with any non-zero exit code, that non-zero
exit code is stored in the hold_subcode.