HTCONDOR-2134 Add job cool down feature
When a job execution attempt is interrupted for any reason and the job
status will return to Idle (or if the starter requested the job be held),
a new cool-down expression (param SYSTEM_ON_VACATE_COOL_DOWN) is
evaluated. If it evaluates to a positive integer, then the job is put into
a cool-down condition (in Idle status)  for that number of seconds. During
this time, the job cannot be executed.
A vacate reason string, code, and subcode are set in the job ad, which can
be referenced by the cool-down expression.

If the starter requested that the job be held, the reason/code/subcode are
used for the vacate attributes. If the cool-down expression doesn't
trigger, then the job is held as in older versions.

Vacate reason codes are in the same namespace as hold reason codes, and
start at value 1000.