HTCONDOR-2389 Robustify job_dagman_rm test
After HTCONDOR-2364 (queued dependent dag job removal)

the existing race condition in this test got worse.  This
test starts a dag, then removes the dag, and verifies the
jobs were removed.  It verifies the removed node jobs by
first condor_wait'ing for them, then checking to see if
the job created a file.  This file is created by the job
after the job polls to see that the dagman proper has
been killed.

We remove that second test in this PR, but as the test still
condor_wait's for the node job to exit, we are still testing
what we say we are.