Add some required attributes into gangliad projection. HTCONDOR-2416
The gangliad may lookup some attributes from ads fetched from the
collector, even if these attributes are not referenced in any metric.
So unconditionally add these attributes into the projection sent
to the collector.

Also for testing purposes, add a bool config knob GANGLIAD_WANT_PROJECTION
to enable or disable projection funcionality.  For now have it default
to False; change to a default of True in an upcoming release
after it runs without incident in devops testing.