HTCONDOR-2415 Blahp runs job in Iwd by default
Add new option blah_run_in_workdir, which disables the blahp's creation
of a temporary subdirectory in which to run the job. This new mode
requires that all directories are shared between submit and worker nodes.
Input files outside the Iwd and remapped output files are still
copied/moved as before.