Implement config knob MONITOR_MULTIPLE_COLLECTORS. HTCONDOR-2417
Normally, the condor_gangliad will only query the default pool collector.
This commit introduced config knob MONITOR_MULTIPLE_COLLECTORS, enabling
the condor_gangliad to monitor a bespoke list of multiple collectors.
OSPool plans to use this to monitor hosted CEs for the CE Dashboard effort.
The implementation is a little obtuse in order to change the execution path
as little as possible in the case that MONITOR_MULTIPLE_COLLECTORS is not
specified (i.e. be conservative, don't break existing deployments).