HTCONDOR-2437 Implementation Pelican local cache
First implentation of the Pelican local cache. This adds the
configuration in both condor_examples as well as a feature template.
While there is stil work to do on both the Pelican and Condor side for
the local cache, the basic functionality is there to get this merged and
begin testing it.