HTCONDOR-2453 Avoid IPv6 link-local network addresses
In some configurations on linux, getaddrinfo() of the local hostname
will return an IPv6 link-local address before any public IPv6 or
IPv4 addresses. This breaks HAD and leads to confusing behavior.
Now, we re-order IPv6 link-local addresses to the end of the list of
addresses in APIs that sit on top of getaddrinfo(). This is in addition
to the optional reordering of IPv4 addresses to the front of the list.