HTCONDOR-2435 Flatten transfer epoch ads
The transfer epoch ads are full of nested ads that, if converted
straight into JSON, would be uselessly impossible to query in Elasticsearch.
Instead, what this aims to do is flatten the ads into multiple JSON docs,
one per file transfer attempt.

Some modification of the attr names was needed to maintain uniqueness
between attempt-based attributes and whole-plugin-based attributes.
For example, for an attempt's transfer time, the end time attr was
converted from TransferEndTime0 to AttemptEndTime (and an Attempt
counter added to the doc).