HTCONDOR-2435 Pass non-default mapping information to interfaces
The mapping for transfer epoch ads does not match that for job ads,
particularly given the flattening required in the previous commit.
This solution of passing the entire ad source object to the
interface object feels a little hacky, but should get the job done.
We should revisit whether there should even be a default mapping
given requests to push arbitrary ad types via the adstash interface.