HTCONDOR-2467 Fix when user credentials are cleaned up by the credd
Remove credential mark files when the user submits a job. This fixes a
race condition where the schedd creates a mark file between when submit
contacts the credd and when it contacts the schedd. As a result, the
credentials are removed while the job is in the queue and possibly
running.

When creating a mark file, don't touch an existing mark file and don't
create one if there are no credentials to clean up. Otherwise,
credentials are never cleaned up and mark files are continually created
for dormant users. This is a consequence of persistent user records in
the schedd.