HTCONDOR-2452 Add new placementd daemon
The placementd manages access to an AP for remote users who don't have
a local OS account. It will assign users to an AP user and OS account,
and issue a placement token (an IDToken) that enables the user to place
jobs at the AP. The new (admin) tool condor_login is used to interact
with the placementd. At present, the only option is to map a new remote
user and obtain a token for them.