HTCONDOR-2523 Remove redundent std::string::c_str() calls when calling into methods that take a std::string.  This avoids a useless copy.
When we still had many MyString's in the code, when we needed to
call a function that took a std::string, but had a MyString in hand,
we'd frequently call .c_str() on the MyString to implicitly
construct the callee's needed std::string.  When we replaced the
MyString's with std::strings, we didn't always remove this now
redundant copy.

clang-tidy does a good job of finding this, but I'm not sure that
each change is guaranteed to be safe.  I've checked each of these,
but we need to make sure the callee is taking a _const_ std::string &,
AND that the caller isn't injecting a null byte in the middle
of the std::string, so that the cstring is shorter than the std::string