Improvements if gangliad fails to contact collector. HTCONDOR-2417
Previously, mapCollectorIPs was called even if we failed to query
the collector being mapped.  Also, previously mapCollectorIps used index == 0
to clear things out, but in the multiple collector case if the first collector pool
was unresponsive we continue resulting the index incrementing and thus never
reseting all the deeper function call attributes.