There are four parts of the protocol to start with:
1.  The starter, receiving the inbox sandbox from the shadow.
2.  The starter, sending the output sandbox to the shadow.
3.  The shadow, sending the input sandbox to the starter.
4.  The shadow, receiving the output sandbox from the starter.

This commit gets (1) to the point where the shadow thinks that file
transfer has succeeded; we can't run a job yet because (a) the starter
was writing files to /dev/null (for simplicity) and (b) the starter
doesn't know that file transfer succeeded (because it depends on
the callbacks for that, which also don't exist yet).

I would like to fix these problems and run a (CEDAR-only) job, and
then try to refactor into something saner, before moving on to 2-4.