HTCONDOR-2584 Fix job submission with idtoken in personal condor
In a personal condor, if idtokens auth is used, condor_submit will create
an ephemeral token with identity condor@password. Job submission will fail,
since 'condor' is not allowed to submit jobs.
Now, the schedd will create a persistent user record for the process owner
of a personal condor and use that as the owner of any jobs submitted by
a 'condor' user.