HTCONDOR-2598 Improve checks of superuser impersonation authorization
Whether a queue superuser may impersonate a user is now based on whether
that user has a user record, not on a seperate list of domain-less
usernames that the schedd has seen go by.
This allows users from foreign domains to successfully run jobs as they
were before (this has its own problems, which will be addressed at a later
time).