HTCONDOR-2601 Improve code to scan local system for network interfaces
When scanning for the local system's network interfaces, keep the binary
format as long as possible (up through sysapi and
network_interface_to_ip/sockaddr()). This reduces the number of
translations between binary and ascii representations.
The main observable benefit is supporting the use of ipv6 link-local
addresses on windows (bringing it up to the flawed level on unix).
It also keeps the os-specific details inside sysapi.