Revert "HTCONDOR-2610 Fix memory leak in HookClientMgr"
This reverts commit cf6429b8030a66d22157ac64fa04ab0f3a51f7dd.

The HookClientMgr memory ownership model is a mess.

Turns out today it's spawn method takes a pointer to a (subclass) of HookClient
and if HookClient->wantsOutput is true, the HookClientMgr takes ownernship of
the hook client object, and deletes it when done with it If HookClient->wantsOutput
is false, HookClientMgr does not take ownership, and does not delete it
The client code knows this, and in most places where the client creates a
HookClient (subclass) with wantOutput = false, it creates the HookClient on the stack.

I don't have time this release to clean this all up, so instead of half
fixing this, I'm going to unfix it now, and we can make another run at
this later.