Turn the interpreter off and back on again on a SIGHUP.  This will have to be handled more carefully when we allow re-entry into the event loop while a generator is still alive, but for now, this allows the Python to change as if it were configuration, which is nice.
More importantly, (eventually!) discover that some of the classad2/common2 util
functions cache their module definitions, which of course means they don't
work if you turn the interpreter off and back on again.  "Fixed", for now, in
a way that they don't cache the module look-ups at all anymore.

Fix up a bunch of (probable?) memory leaks.  (For production, we'd still
need to run the interpreter in leak-detection mode.)