Fix repeated non-fungible machine resources. HTCONDOR-2678
An earlier commit removed usage of the StringList class from the condor_startd,
and for MACHINE_RESOURCES, replaced the StringList object with an STL std::set object.
However, a StringList allows repeated elements, while a std::set does not, resulting in
repeating non-fungible resources (like repeated GPU ids which trying to
run multiple jobs on one GPU) broken.
In this patch, we simply replace the std::set with a std::multiset
(which allows repeats), as ordering does not matter in this case.