HTCONDOR-2408 Add condor-credmon-multi RPM package
The condor-credmon-multi package is an alternative to the
condor-credmon-vault package that doesn't clobber the config.
Thus, Vault should work alongside the Local and OAuth credmons
installed via the condor-credmon-local and condor-credmon-oauth
packages now that the condor_credmon_oauth "daemon" is able to
run all three concurrently.

Also, I noticed that urllib3 is used by the VaultCredmon class
without ever explicitly requiring it before, that is now fixed.

Also, the condor-credmon-local package (the base for both
condor-credmon-oauth and condor-credmon-multi packages) is now
explicitly marked as conflicting with condor-credmon-vault.