Preserve ordering of non-fungible resources in startd. HTCONDOR-2684
While it is unlikely to matter matter for GPU management, it is possible that some sites
may assume that the order in which non-fungible resources are defined in the config file
is the order in which that are assigned to slots (since was true HTCSS 23.0).
To restore the validity of this assumption, the data structure in the startd needs to
change from a std:multiset  to std::vector (which can preserve the order elements are added).