HTCONDOR-2782 Use the best stat() calls on all platforms and simplify code
On the unix side, all of our platforms are 64-bit and we don't need to
mess with stat() vs stat64() anymore.
On the windows side, we ditch the DST workaround, which shouldn't be
necessary anymore and was only used in some locations. Also, use _stat64()
and friends, which always use 64-bit filesize and timestamp values.

We keep the #define of stat() and company on windows, since a lot of code
uses that. We also add condor_stat() and company, which makes it more
obvious that that actual OS function isn't necessarily named stat().