HTCONDOR-2793 Blahp now merges job and worker node PATHs
When the job description given to the blahp includes a PATH environment
variable, the blahp now concatenates it with the PATH that's already set
on the worker node (job PATH comes first).
This can be disabled (so that only the job-provided PATH value is used)
by setting blah_merge_paths=no in the blah.config file.