HTCONDOR-2833 Add capability form of IDTokens
Add a new form of IDToken that acts as a capability. The jwt has no subject
or scope claims. These values are only stored in the tokens database by
the creator of the token. During CEDAR authentication, the server uses the
jti claim to lookup the token in its database and retrieve the subject and
scope claims from there.
As with regular IDTokens, the scope constrains the authorizations for the
client from those granted by the server's configuration.
This means that a capability token can only be used to authenticate with a
daemon that's part of the same HTCondor installation as the creator of the
token (absent some mechanism to diseminate token information across nodes
of an HTCondor system).